+++ NSCollection_utils.h
@@ -58,6 +58,6 @@
 @interface NSMutableArray (Sorting)
 
 - (void)sortUnstableUsingFunction:(NSInteger (*)(id *, id *))compare;
-- (void)sortStableUsingFunction:(NSInteger (*)(id *, id *))compare usingBuffer:(id **)buffer ofSize:(unsigned int*)bufSize;
+- (void)sortStableUsingFunction:(NSInteger ( *)(id *, id *))compare usingBuffer:(id **)buffer ofSize:(NSUInteger *)bufSize;
 @end
